requirements.txt
appnope==0.1.2
asgiref==3.2.10
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
Django==3.1
django-3-jet==1.0.8
django-common-helpers==0.9.2
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
logger==1.4
MarkupSafe==1.1.1
oauthlib==3.1.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.5
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.3.1
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
